% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_setup.R
\name{csv2vec}
\alias{csv2vec}
\title{loading parameters from a csv file}
\usage{
csv2vec(pathCsv, iLine)
}
\arguments{
\item{pathCsv}{path to csv file}

\item{iLine}{index corrsponding to the line to read}
}
\value{
a named vector containing parameters of configuration setup
}
\description{
loading parameters from a csv file
}

