% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_fields}
\alias{generate_fields}
\title{generate fields from parameters}
\usage{
generate_fields(vect_param, nb_fields = 100, verbose = F)
}
\arguments{
\item{vect_param}{a vector containing parameters for fields generation}

\item{nb_fields}{an integer specifying the number of fields to generate}

\item{verbose}{boolean for whether to display comments}
}
\value{
a dataframe where each column corresponds to generated field
}
\description{
generate fields from parameters
}

