% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_meas}
\alias{generate_meas}
\title{sample measurement from generated fields}
\usage{
generate_meas(vect_meas, Y_xy, verbose = F)
}
\arguments{
\item{Y_xy}{original fields}

\item{verbose}{boolean for whether to display comments}

\item{vect_param}{a vector containing parameters for sampling measurements}
}
\value{
a dataframe containing measurements
}
\description{
sample measurement from generated fields
}

