% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_fields}
\alias{names_fields}
\title{names of parameters for fields}
\format{a vector of characters}
\source{
self-defined
}
\usage{
names_fields
}
\description{
A vector containing the names of parameters
 used when generating fields
}
\keyword{datasets}

