% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_meas}
\alias{names_meas}
\title{A vector containing the names of parameters
 used when sampling measurements from fields}
\format{a vector of characters}
\source{
self-defined
}
\usage{
names_meas
}
\description{
A vector containing the names of parameters
 used when sampling measurements from fields
}
\keyword{datasets}

